phase 1 done  feat: enhance comment module with delete functionality and improve error handling

+++ src/utils/cloudinary.helper.ts
@@ -5,11 +5,12 @@ export const uploadToCloudinary = async (localFilePath: string) => {
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "image",
     });
-    fs.unlinkSync(localFilePath);
+
     return response.secure_url;
-  } catch (error) {
-    console.log("Error while uploading the image to cloudinary", error);
-    fs.unlinkSync(localFilePath);
+  } finally {
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
   }
 };
